fixed servo pwm
the problem was that some functions was following the fixed 100 steps not the variable speed_steps

+++ SOFTWARE/FAC firmware/Core/Inc/servo.h
@@ -11,7 +11,7 @@
 #include "main.h"
 
 void setServoIsEnable(uint8_t state, TIM_HandleTypeDef *htim);
-void setServoAngle(uint8_t servoN, uint8_t percentage);
+void setServoAngle(uint8_t servoN, uint16_t percentage);
 
 enum SERVO{ S1, S2};
